Refactor: moving structs 구조체 타입 파일 옮김
-
